+++ compiler/x/amd64/codegen/AMD64SystemLinkage.cpp
@@ -901,7 +901,7 @@ TR::AMD64ABILinkage::mapIncomingParms(
          uint32_t align = getAlignment(parmCursor->getDataType());
          uint32_t alignMinus1 = (align <= AMD64_STACK_SLOT_SIZE) ? (AMD64_STACK_SLOT_SIZE - 1) : (align - 1);
          uint32_t pos = -stackIndex;
-         pos += parmCursor->getSize();
+         pos += static_cast<uint32_t>(parmCursor->getSize());
          pos = (pos + alignMinus1) & (~alignMinus1);
          stackIndex = -pos;
          parmCursor->setParameterOffset(stackIndex);
